requirements.txt
opencv-python==4.12.0.88
mediapipe==0.10.21
numpy==1.26.4
pyttsx3==2.99
scikit-learn==1.7.2
Pillow==12.0.0
